[refactor] Refactored the packagestructure, so the delegate can be mostly package private
